+++ R/reexport.R
@@ -6,3 +6,12 @@ bbotk::trm
 
 #' @export
 bbotk::trms
+
+#' @export
+mlr3misc::mlr_callbacks
+
+#' @export
+mlr3misc::clbk
+
+#' @export
+mlr3misc::clbks