+++ R/zzz.R
@@ -12,8 +12,10 @@
   x = utils::getFromNamespace("mlr_reflections", ns = "mlr3")
   x$tuner_properties = "dependencies"
 
-  x = utils::getFromNamespace("bbotk_reflections", ns = "bbotk")
-  x$callback_steps = c(x$callback_steps, "on_eval_after_design", "on_eval_after_benchmark", "on_eval_after_aggregation")
+  # callbacks
+  x = utils::getFromNamespace("mlr_callbacks", ns = "mlr3misc")
+  x$add("mlr3tuning.early_stopping", load_callback_early_stopping)
+  x$add("mlr3tuning.backup", load_callback_backup)
 
   assign("lg", lgr::get_logger("bbotk"), envir = parent.env(environment()))
   if (Sys.getenv("IN_PKGDOWN") == "true") {