+++ R/TuningInstanceMulticrit.R
@@ -3,8 +3,8 @@
 #' @include TuningInstanceSingleCrit.R
 #'
 #' @description
-#' The [TuningInstanceMultiCrit] specifies a tuning task for [Tuners][Tuner].
-#' The instance is not created by the user but internally when the function [tune()] is called with more than one [mlr3::Measure].
+#' The [TuningInstanceMultiCrit] specifies a tuning problem for [Tuners][Tuner].
+#' The function [ti()] creates a [TuningInstanceMultiCrit] and the function [tune()] creates an instance internally.
 #'
 #' @inherit TuningInstanceSingleCrit details
 #' @inheritSection TuningInstanceSingleCrit Resources
@@ -27,17 +27,25 @@
 #'
 #' @export
 #' @examples
+#' # get learner and define search space
 #' learner = lrn("classif.rpart", cp = to_tune(1e-04, 1e-1, logscale = TRUE))
 #'
-#' instance = tune(
-#'   method = "random_search",
+#' # construct tuning instance
+#' instance = ti(
 #'   task = tsk("pima"),
 #'   learner = learner,
 #'   resampling = rsmp ("holdout"),
-#'   measures = msrs(c("classif.ce", "classif.acc")),
-#'   term_evals = 4)
+#'   measures = msrs(c("classif.ce", "time_train")),
+#'   terminator = trm("run_time", secs = 10)
+#' )
 #'
-#' # get optimized hyperparameters
+#' # get tuner
+#' tuner = tnr("random_search", batch_size = 10)
+#'
+#' # tune classification tree on pima data set
+#' tuner$optimize(instance)
+#'
+#' # get result
 #' instance$result
 TuningInstanceMultiCrit = R6Class("TuningInstanceMultiCrit",
   inherit = OptimInstanceMultiCrit,
@@ -45,10 +53,6 @@ TuningInstanceMultiCrit = R6Class("TuningInstanceMultiCrit",
 
     #' @description
     #' Creates a new instance of this [R6][R6::R6Class] class.
-    #'
-    #' This defines the resampled performance of a learner on a task, a
-    #' feasibility region for the parameters the tuner is supposed to optimize,
-    #' and a termination criterion.
     initialize = function(task, learner, resampling, measures, terminator, search_space = NULL, store_benchmark_result = TRUE, store_models = FALSE, check_values = FALSE, allow_hotstart = FALSE, keep_hotstart_stack = FALSE, evaluate_default = FALSE, callbacks = list()) {
       private$.evaluate_default = assert_flag(evaluate_default)
       learner = assert_learner(as_learner(learner, clone = TRUE))
@@ -77,8 +81,8 @@ TuningInstanceMultiCrit = R6Class("TuningInstanceMultiCrit",
     },
 
     #' @description
-    #' The [Tuner] object writes the best found points
-    #' and estimated performance values here. For internal use.
+    #' The [Tuner] object writes the best found points and estimated performance values here.
+    #' For internal use.
     #'
     #' @param ydt (`data.table::data.table()`)\cr
     #'   Optimal outcomes, e.g. the Pareto front.