+++ NEWS.md
@@ -1,5 +1,7 @@
 # mlr3tuning 0.14.0.9000
 
+* feat: Add `ti()` function to create a `TuningInstanceSingleCrit` or `TuningInstanceMultiCrit`.
+
 # mlr3tuning 0.14.0
 
 * feat: Add option `evaluate_default` to evaluate learners with hyperparameters set to their default values.