+++ NEWS.md
@@ -1,4 +1,4 @@
-# mlr3tuning 0.14.0.9000
+# mlr3tuning 0.15.0
 
 * fix: `AutoTuner` has not accepted `TuningSpace` objects as search spaces.
 * feat: Add `ti()` function to create a `TuningInstanceSingleCrit` or `TuningInstanceMultiCrit`.